Rewrite core API Gateway to Go HTTP translator
Usage shows that inner behaviour of Go HTTP listeners are far more
complex than the first approach we used to translate API Gateway
requests. This new refactoring removes custom listeners and connections
in favour of a more standard approach. A real proxy is created and it
uses official Go TCP listener on local loopback. Each time a new request
comes from API Gateway, it is translated to a regular HTTP request and
standard Go HTTP client is used to forward the request to the previously
created proxy.

This new version is backward compatible unless one has used the
low-level listeners and connections outside the apigatewayproxy package.

Special thanks to @ludovicbailly (twitter), for his usecase with Go
reverse proxies, which showed the weakness of the previous version of
this lib.

CAT: #feat #break #doc