Handle response errors more properly
Value returned by http client Do function is nil or already closed if
there is any error. See
https://golang.org/src/net/http/client.go?s=16518:16570#L480
Therefore in order to avoid memory leaks in a potential ioutil.ReadAll
error and to be Go idioms compliant, the response body is deferred to be
closed immediately after checking there is no error with the Do function.

CAT: #fix
REF: #8
THX: @Zambiorix